Cat_Pre_App/requirements.txt
sys
streamlit
streamlit.web.cli
joblib
pandas
os
numpy
sklearn.metrics
sklearn.preprocessing
matplotlib.pyplot
seaborn
rdkit
rdkit.Chem.Draw
base64
io
itertools
sklearn.ensemble
sklearn.preprocessing
sklearn.model_selection